Moved to dto.pure all the only python parts of the DTO, thus separating all the data from the functionality, as it should be
